requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.1.1
requests==2.32.5
simple-term-menu==1.6.6
urllib3==2.5.0
